requirements.txt
fastApi
uvicorn
sqlalchemy
pymysql
alembic
jinja2
sqlalchemy_imageattach
Wanda
autopep8
pydantic[email]
passlib[bcrypt]
starlette
bcrypt
python-multipart
python-jose[cryptography]